+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * *BREAKING* Migrate stripes dependencies to their Sunflower versions. Refs UIF-574.
 * *BREAKING* Migrate `react-intl` to v7. Refs UIF-575.
 * Batch allocation bugs/enhancements. Refs UIF-589.
+* Prevent sorting when data entered in batch fund allocation sceen. Refs UIF-570.
 
 ## [7.0.1](https://github.com/folio-org/ui-finance/tree/v7.0.1) (2024-12-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v7.0.0...v7.0.1)